+++ src/onegov.core/onegov/core/tests/test_cronjobs.py
@@ -42,6 +42,17 @@ def test_is_scheduled_at():
     assert job.is_scheduled_at(datetime(2015, 1, 1, 7, 0, 1), 'UTC')
     assert job.is_scheduled_at(datetime(2015, 1, 1, 7, 0, 59), 'UTC')
 
+    # hours can be wildcards
+    job = cronjobs.Job(lambda: None, hour='*', minute=0, timezone='CET')
+    assert job.is_scheduled_at(datetime(2015, 1, 1, 8, 0), 'UTC')
+    assert job.is_scheduled_at(datetime(2015, 1, 1, 9, 0), 'UTC')
+    assert not job.is_scheduled_at(datetime(2015, 1, 1, 9, 15), 'UTC')
+
+    job = cronjobs.Job(lambda: None, hour='*', minute=15, timezone='CET')
+    assert not job.is_scheduled_at(datetime(2015, 1, 1, 8, 0), 'UTC')
+    assert not job.is_scheduled_at(datetime(2015, 1, 1, 9, 0), 'UTC')
+    assert job.is_scheduled_at(datetime(2015, 1, 1, 9, 15), 'UTC')
+
 
 def test_overlapping_cronjobs():
 
@@ -60,6 +71,42 @@ def test_overlapping_cronjobs():
 
     with pytest.raises(ConflictError):
         App.commit()
+
+
+def test_overlapping_cronjobs_with_wildcards():
+
+    class App(Framework):
+        pass
+
+    @App.cronjob(hour='*', minute=0, timezone='UTC')
+    def first_job(request):
+        pass
+
+    @App.cronjob(hour=5, minute=0, timezone='UTC')
+    def second_job(request):
+        pass
+
+    scan_morepath_modules(App)
+
+    with pytest.raises(ConflictError):
+        App.commit()
+
+
+def test_overlapping_cronjobs_with_wildcards_no_conflict():
+
+    class App(Framework):
+        pass
+
+    @App.cronjob(hour='*', minute=0, timezone='UTC')
+    def first_job(request):
+        pass
+
+    @App.cronjob(hour=5, minute=15, timezone='UTC')
+    def second_job(request):
+        pass
+
+    scan_morepath_modules(App)
+    App.commit()
 
 
 def test_non_5_minutes_cronjobs():